ピクセル生産シーン（ピクセルカラー指定のRGB値指定は完成、ピクセル生産部分着手、rgbが0のときに進捗しない問題がある）,13

+++ Idle Color sRpG Project/Assets/Scripts/SaveClass.cs
@@ -391,7 +391,7 @@ public class SaveClass// : MonoBehaviour
 
                         if (values[0].Equals("CharactersAll[" + i.ToString() + "].Stats[" + j.ToString() + "].SPD"))
                         {
-                            CharactersAll[i].Stats[j].SPD = (uint)(int.Parse(values[1]));
+                            CharactersAll[i].Stats[j].SPD = (byte)(int.Parse(values[1]));
                             FlagHit = true;
                             break;
                         }
